+++ fcxref/rename/__init__.py
@@ -1,4 +1,4 @@
-from .make_rename import make_rename
+from .make_rename_property import make_rename_property
 
 
-__all__ = ['make_rename']
+__all__ = ['make_rename_property']